Update to Kotlin 2.4.20, JUnit 5.14.4, mockk 1.14.11, Jackson 2.21.5 (#2449)
* Update Kotlin to 2.4.20, JUnit to 5.14.4, mockk to 1.14.11, Jackson to 2.21.5.

Give processReportBlock an explicit return type: an expression-bodied
function containing return statements becomes an error in Kotlin 2.5
(KTLC-288). Jackson 2.21.5 matches the version jitsi-utils now uses.

* Remove redundancies flagged by Kotlin 2.4.

Drop redundant toLong()/toInt()/toDouble() conversions, elvis operators
whose right-hand side is null or can never be reached, an always-true null
check, a redundant else branch in an exhaustive when, an unnecessary !!,
and use Any() rather than java.lang.Object() for a lock.

* Update jitsi-utils to 1.0-153, jicoco to 1.1-179, jitsi-xmpp-extensions to 1.0-124, jitsi-metaconfig to 1.0-18, ice4j to 3.2-18.

All built with Kotlin 2.4.20. The metaconfig bump is required alongside
jicoco 1.1-179, which calls ConfigSource.getDescription(), absent from the
1.0-11 previously pinned here.

+++ jitsi-media-transform/src/main/kotlin/org/jitsi/nlj/rtp/bandwidthestimation/GoogleCcEstimator.kt
@@ -61,7 +61,7 @@ class GoogleCcEstimator(diagnosticContext: DiagnosticContext, parentLogger: Logg
      * Implements the loss-based part of Google CC.
      */
     private val sendSideBandwidthEstimation =
-        SendSideBandwidthEstimation(diagnosticContext, initBw.bps.toLong(), logger).also {
+        SendSideBandwidthEstimation(diagnosticContext, initBw.bps, logger).also {
             it.setMinMaxBitrate(minBw.bps.toInt(), maxBw.bps.toInt())
         }
